node/requirements.txt
python-multipart
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0

asyncpg==0.22.0
psycopg2-binary==2.8.6
aiofiles

requests
PyYAML
humanfriendly